Moved renderer in a separate class
